+++ src/linktransformer/__init__.py
@@ -1,7 +1,8 @@
 # __init__.py
 
-__version__ = "0.0.1"
+__version__ = "0.1.1"
 __MODEL_HUB_ORGANIZATION__ = 'sentence-transformers'
+from .data import DATA_DIR_PATH
 from .infer import *
 from .preprocess import *
 from .train_model import *