+++ tests/train_test.py
@@ -1,14 +1,13 @@
 import os
-from linktransformer.train_model import train_model
+import linktransformer as lt
 from linktransformer.data import DATA_DIR_PATH
-from linktransformer.configs import LINKAGE_CONFIG_PATH
 
 def test_train_model():
     # Define the path to the test dataset
     dataset_path = os.path.join(DATA_DIR_PATH, "es_mexican_products.xlsx")
 
     # Call the train_model function
-    saved_model_path = train_model(
+    saved_model_path = lt.train_model(
         model_path="hiiamsid/sentence_similarity_spanish_es",
         data=dataset_path,
         left_col_names=["description47"],
@@ -16,9 +15,10 @@ def test_train_model():
         left_id_name=['tariffcode47'],
         right_id_name=['tariffcode48'],
         log_wandb=False,
-        config_path=LINKAGE_CONFIG_PATH,
-        training_args={"num_epochs": 3,
-                       "test_at_end": False}
+        training_args={"num_epochs": 0,
+                       "test_at_end": False,
+                       "save_val_test_pickles": True,
+                       "model_save_name": "test_model"}
     )
 
     # Add assertions to check if the training was successful and the model was saved